+++ src/components/pages/Earn/PoolsOverview.tsx
@@ -12,7 +12,7 @@ import { PLATFORM_METADATA } from './constants';
 import { TokenIconSvg } from '../../icons/TokenIcon';
 import { EtherscanLink } from '../../core/EtherscanLink';
 import { ExternalLink } from '../../core/ExternalLink';
-import { AccentColors } from '../../../types';
+import { AccentColors, Platforms } from '../../../types';
 import { Tooltip } from '../../core/ReactTooltip';
 
 const ApyAmount = styled(Amount)`
@@ -32,7 +32,7 @@ const TableGroup = styled.div`
 `;
 
 const PlatformIcon = styled(TokenIconSvg)`
-  margin-right: 8px;
+  margin-right: 16px;
 `;
 
 const PlatformContainer = styled.div<{ colors: AccentColors }>`
@@ -97,7 +97,15 @@ export const PoolsOverview: FC<{}> = () => {
 
   const [activePools, otherPools, expiredPools] = useMemo(() => {
     const items = Object.values(stakingRewardsContracts)
-      .sort()
+      .sort(a => {
+        // *teleports behind you*
+        return a.pool.platform === Platforms.Curve
+          ? -1
+          : // nothing personnel, kid
+          a.pool.platform === Platforms.Balancer
+          ? 0
+          : 1;
+      })
       .map(item => {
         const {
           address: id,
@@ -157,13 +165,6 @@ export const PoolsOverview: FC<{}> = () => {
                       )}
                     </>
                   );
-                // case Columns.StakingApy:
-                //   return (
-                //     <ApyAmount
-                //       amount={stakingRewardsContract.stakingTokenApy}
-                //       format={NumberFormat.Percentage}
-                //     />
-                //   );
                 case Columns.RewardsApy: {
                   if (expired) {
                     return <>N/A</>;
@@ -174,7 +175,11 @@ export const PoolsOverview: FC<{}> = () => {
                         amount={item.apy.value}
                         format={NumberFormat.CountupPercentage}
                       />
-                      {item.platformRewards ? <ApyNote>+ BAL</ApyNote> : null}
+                      {item.platformRewards ? (
+                        <ApyNote>
+                          + {item.platformRewards.platformToken.symbol}
+                        </ApyNote>
+                      ) : null}
                     </div>
                   ) : item.apy.waitingForData ? (
                     <Tooltip tip="Calculating APY requires data from 24h ago, which is not available yet.">
@@ -197,9 +202,14 @@ export const PoolsOverview: FC<{}> = () => {
                         price={rewardsToken.price}
                       />
                       {platformRewards ? (
-                        <Tooltip tip="Currently BAL rewards are airdropped based on Balancer's reward programme allocations. Earned rewards can be claimed on the EARN dashboard.">
+                        <Tooltip
+                          tip={
+                            item.curve
+                              ? 'This pool receives CRV rewards'
+                              : "Currently BAL rewards are airdropped based on Balancer's reward programme allocations. Earned rewards can be claimed on the EARN dashboard."
+                          }
+                        >
                           <TokenAmount
-                            // amount={platformRewards.totalPlatformRewards}
                             format={NumberFormat.Abbreviated}
                             symbol={platformRewards.platformToken.symbol}
                             price={platformRewards.platformToken.price}