+++ src/context/earn/EarnDataProvider.tsx
@@ -25,6 +25,7 @@ export const EarnDataProvider: FC<{}> = ({ children }) => {
 
   // Data that we subscribe to on each block:
   // - Staking rewards contracts
+  // - Curve pool balances (if available)
   const rawEarnData = useRawEarnData(syncedEarnData);
 
   const earnData = useMemo(