+++ src/types.ts
@@ -10,6 +10,8 @@ import { StakingRewardsWithPlatformToken as IStakingRewardsWithPlatformToken } f
 import { BigDecimal } from './web3/BigDecimal';
 import { RewardsDistributor as IRewardsDistributor } from './typechain/RewardsDistributor.d';
 import { MerkleDrop as IMerkleDrop } from './typechain/MerkleDrop.d';
+import { MusdGauge } from './typechain/MusdGauge.d';
+import { TokenMinter as TokenMinterImport } from './typechain/TokenMinter.d';
 
 export interface Transaction {
   formId?: string;
@@ -54,6 +56,8 @@ export enum Interfaces {
   StakingRewardsWithPlatformToken,
   RewardsDistibutor,
   MerkleDrop,
+  Gauge,
+  TokenMinter,
 }
 
 export interface Instances {
@@ -64,6 +68,8 @@ export interface Instances {
   [Interfaces.StakingRewardsWithPlatformToken]: IStakingRewardsWithPlatformToken;
   [Interfaces.RewardsDistibutor]: IRewardsDistributor;
   [Interfaces.MerkleDrop]: IMerkleDrop;
+  [Interfaces.Gauge]: MusdGauge;
+  [Interfaces.TokenMinter]: TokenMinterImport;
 }
 
 /**
@@ -144,6 +150,7 @@ export interface Connector {
 export enum Platforms {
   Balancer = 'Balancer',
   Uniswap = 'Uniswap',
+  Curve = 'Curve',
 }
 
 export interface BlockTimestamp {